actor: add deduplication context
